Decouple CoreAppSetup from AppSetup
